+++ src/lib.rs
@@ -1,4 +1,6 @@
 #[cfg(feature = "toml")]
+pub mod cargo_merge;
+#[cfg(feature = "toml")]
 pub mod cargo_pre_ci;
 #[cfg(feature = "toml")]
 pub mod pyproject_merge;